♻️ Refactor: 스케줄 캘린더 헤더 로직 분리 (#89)
